requirements.txt
fastapi
uvicorn[standard]
pymupdf
pillow
pytesseract
chromadb
openai
tiktoken
python-multipart
python-dotenv

